Update with apollo
HttpLinkModule It’s optional, you can replace it with any other Link. Its biggest advantage of all is that it uses HttpClient internally so it’s possible to use it in NativeScript or in combination with any other HttpClient provider. By using HttpLinkModule you get Server-Side Rendering for free, without any additional work.